backend/requirements.txt
llama-index==0.10.23
fastapi==0.109.0
python-dotenv